refactor: extract rate-limit scheduler
