+++ open-sse/services/rateLimitManager.ts
@@ -26,6 +26,7 @@ import {
   parseResetTime,
   toPlainHeaders,
 } from "./rateLimitManager/headers";
+import { scheduleWithAbort } from "./rateLimitManager/scheduler.ts";
 
 interface LearnedLimitEntry {
   provider: string;
@@ -512,42 +513,6 @@ function getLimiter(provider: string, connectionId: string, model: string | null
 
   limiterLastUsed.set(key, Date.now());
   return limiter;
-}
-
-async function scheduleWithAbort<T>(
-  limiter: Bottleneck,
-  scheduleOpts: { expiration?: number },
-  fn: () => T | Promise<T>,
-  signal: AbortSignal | null,
-): Promise<T> {
-  if (!signal) return limiter.schedule(scheduleOpts, () => Promise.resolve(fn()));
-
-  let abortListener: (() => void) | undefined;
-  const abortPromise = new Promise<never>((_, reject) => {
-    const onAbort = () => {
-      const reason = signal.reason;
-      const err = reason instanceof Error
-        ? reason
-        : new Error(typeof reason === "string" ? reason : "The operation was aborted");
-      err.name = "AbortError";
-      reject(err);
-    };
-    if (signal.aborted) {
-      onAbort();
-      return;
-    }
-    abortListener = onAbort;
-    signal.addEventListener("abort", abortListener, { once: true });
-  });
-
-  try {
-    return await Promise.race([
-      limiter.schedule(scheduleOpts, () => Promise.resolve(fn())),
-      abortPromise,
-    ]);
-  } finally {
-    if (abortListener) signal.removeEventListener("abort", abortListener);
-  }
 }
 
 /**
