+++ open-sse/services/rateLimitManager/scheduler.ts
@@ -1,0 +1,40 @@
+import Bottleneck from "bottleneck";
+
+type ScheduleOptions = { expiration?: number };
+
+/** Schedule a request while allowing an AbortSignal to cancel queued work. */
+export async function scheduleWithAbort<T>(
+  limiter: Bottleneck,
+  scheduleOpts: ScheduleOptions,
+  fn: () => T | Promise<T>,
+  signal: AbortSignal | null,
+): Promise<T> {
+  if (!signal) return limiter.schedule(scheduleOpts, () => Promise.resolve(fn()));
+
+  let abortListener: (() => void) | undefined;
+  const abortPromise = new Promise<never>((_, reject) => {
+    const onAbort = () => {
+      const reason = signal.reason;
+      const err = reason instanceof Error
+        ? reason
+        : new Error(typeof reason === "string" ? reason : "The operation was aborted");
+      err.name = "AbortError";
+      reject(err);
+    };
+    if (signal.aborted) {
+      onAbort();
+      return;
+    }
+    abortListener = onAbort;
+    signal.addEventListener("abort", abortListener, { once: true });
+  });
+
+  try {
+    return await Promise.race([
+      limiter.schedule(scheduleOpts, () => Promise.resolve(fn())),
+      abortPromise,
+    ]);
+  } finally {
+    if (abortListener) signal.removeEventListener("abort", abortListener);
+  }
+}
